requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
pytest
respx
pytest-asyncio